Split behaviors into separate files.
