+++ src/engine/ai/behavior-wander.ts
@@ -1,0 +1,1 @@
+export const wanderBetweenRooms = 5
